cap nhat backend + ma hoa password
cap nhat backend + ma hoa password

+++ src/main/java/com/poly/assignment/dto/ThongKeDoanhThuLoaiHangDTO.java
@@ -1,13 +1,10 @@
 package com.poly.assignment.dto;
 
-import java.math.BigDecimal;
-
-
 public interface ThongKeDoanhThuLoaiHangDTO {
     String getTenLoai();
-    BigDecimal getTongDoanhThu();
-    Long getTongSoLuong();
-    BigDecimal getGiaCaoNhat();
-    BigDecimal getGiaThapNhat();
-    BigDecimal getGiaTrungBinh();
+    Double getTongDoanhThu();
+    Integer getTongSoLuong();
+    Double getGiaCaoNhat();
+    Double getGiaThapNhat();
+    Double getGiaTrungBinh();
 }
